WIP: More debugging.
Currently nodes seem to crash sometimes at what seems to be right before
the end of a sync phase.
This seem to happen *rhoughly* once every 10min but not in a predictable
manner.
Also the tree construction does not seem to react as graceful to this as it
could be and builds distance-escalating cycles until the maximum distance is
reached.